Fixes gub #61 -> Homepage not displayed if no questions available

+++ grails-app/controllers/au/org/ala/taxonoverflow/AdminController.groovy
@@ -27,19 +27,25 @@ class AdminController {
 
     def ajaxReindexAll() {
 
-        elasticSearchService.deleteAllQuestionsFromIndex()
-        def c = Question.createCriteria()
-        def questionIds = c.list {
-            projections {
-                property("id")
+        if (Question.count > 0) {
+            elasticSearchService.deleteAllQuestionsFromIndex()
+            def c = Question.createCriteria()
+            def questionIds = c.list {
+                projections {
+                    property("id")
+                }
             }
+
+            questionIds?.each { questionId ->
+                elasticSearchService.scheduleQuestionIndexation(questionId)
+            }
+
+            renderResults([success: true, questionCount: questionIds?.size() ?: 0])
+        } else {
+            renderResults([success: true, questionCount: 0])
         }
 
-        questionIds?.each { questionId ->
-            elasticSearchService.scheduleQuestionIndexation(questionId)
-        }
 
-        renderResults([success: true, questionCount: questionIds?.size() ?: 0])
     }
 
     private renderResults(Object results, int responseCode = 200) {
